LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/element-request-fullscreen-dialog-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: dialog.requestFullscreen is not a function. (In 'dialog.requestFullscreen()', 'dialog.requestFullscreen' is undefined)

Harness Error (FAIL), message = Unhandled rejection: dialog.requestFullscreen is not a function. (In 'dialog.requestFullscreen()', 'dialog.requestFullscreen' is undefined)

TIMEOUT Element#requestFullscreen() for dialog element Test timed out

Harness Error (FAIL), message = Unhandled rejection: dialog.requestFullscreen is not a function. (In 'dialog.requestFullscreen()', 'dialog.requestFullscreen' is undefined)

TIMEOUT Element#requestFullscreen() for dialog element Test timed out
